+++ run_dev.sh
@@ -25,7 +25,7 @@ backup_pools:
     rsync_module: docker-backup
     remote_host: 10.0.13.21:30026
 
-staging_dir: /Users/zero/Files/Repos/_temp/staging
+tmp_dir: /Users/zero/Files/Repos/_temp/tmp
 
 web_ui:
   port: 8000